python/requirements.txt
# Python dependencies for Snapchat Memories scripts
requests
pillow
